backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
pydantic==2.13.3
pydantic-settings==2.9.1
python-multipart==0.0.20
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pypdf==5.4.0
python-docx==1.1.2
requests==2.32.3
google-auth==2.40.1
psycopg2-binary==2.9.10
lxml==6.1.0
